Creating list component and filtering tasks
Component to render lists created and I added a method to filter tasks according to the list they belong
